Moved firebase setup from MainContainer to App.js and separated out the config
