+++ src/app/components/Main/MainContainer.js
@@ -5,13 +5,6 @@ import { connect } from 'react-redux';
 import { setAuthState } from 'redux/actions.js';
 import Main from './Main.js';
 
-var config = {
-    apiKey: "AIzaSyCtMk74zXBsAYRYIzamcyRXyGDFP3vKXhA",
-    authDomain: "testing-project-a5d42.firebaseapp.com",
-    databaseURL: "https://testing-project-a5d42.firebaseio.com/"
-};
-firebase.initializeApp(config);
-
 class MainContainer extends React.Component {
     constructor(props) {
         super(props);
