+++ src/app/firebaseConfig.js
@@ -1,0 +1,7 @@
+var config = {
+    apiKey: "AIzaSyCtMk74zXBsAYRYIzamcyRXyGDFP3vKXhA",
+    authDomain: "testing-project-a5d42.firebaseapp.com",
+    databaseURL: "https://testing-project-a5d42.firebaseio.com/"
+};
+
+export default config;
